Visual still stutters a bit

+++ HW6 - Easy Animator View/src/model/Keyframe.java
@@ -6,25 +6,26 @@ package model;
 public class Keyframe {
 
   private String name;
-  private int t1;
-  private int x1;
-  private int y1;
-  private int w1;
-  private int h1;
-  private int r1;
-  private int g1;
-  private int b1;
-  private int t2;
-  private int x2;
-  private int y2;
-  private int w2;
-  private int h2;
-  private int r2;
-  private int g2;
-  private int b2;
+  private double t1;
+  private double x1;
+  private double y1;
+  private double w1;
+  private double h1;
+  private double r1;
+  private double g1;
+  private double b1;
+  private double t2;
+  private double x2;
+  private double y2;
+  private double w2;
+  private double h2;
+  private double r2;
+  private double g2;
+  private double b2;
 
   /**
    * Constructs a Keyframe that has motion.
+   *
    * @param name is the name of shape that will move
    * @param t1 is the start time of this motion
    * @param x1 is the start x value of this shape
@@ -43,9 +44,9 @@ public class Keyframe {
    * @param g2 is the end green value of this shape
    * @param b2 is the end blue value of this shape
    */
-  Keyframe(String name, int t1, int x1, int y1, int w1, int h1, int r1, int g1, int b1,
-      int t2,
-      int x2, int y2, int w2, int h2, int r2, int g2, int b2) {
+  Keyframe(String name, double t1, double x1, double y1, double w1, double h1, double r1, double g1,
+      double b1, double t2, double x2, double y2, double w2, double h2, double r2, double g2,
+      double b2) {
     this.name = name;
     this.t1 = t1;
     this.x1 = x1;
@@ -68,6 +69,7 @@ public class Keyframe {
   /**
    * This constructor is used when setting up the inital state of a shape. The start and end states
    * of the shape will be the same.
+   *
    * @param name is the name ofthe shape
    * @param t1 is the time this shape will appear
    * @param x1 is the starting x position of the shape
@@ -78,7 +80,8 @@ public class Keyframe {
    * @param g1 is the starting green value of the shape
    * @param b1 is the starting blue value of the shape
    */
-  Keyframe(String name, int t1, int x1, int y1, int w1, int h1, int r1, int g1, int b1) {
+  Keyframe(String name, double t1, double x1, double y1, double w1, double h1, double r1, double g1,
+      double b1) {
     this.name = name;
     this.t1 = t1;
     this.x1 = x1;
@@ -100,6 +103,7 @@ public class Keyframe {
 
   /**
    * Returns the name of the shape of this keyframe.
+   *
    * @return the name.
    */
   public String getName() {
@@ -108,129 +112,147 @@ public class Keyframe {
 
   /**
    * Gets the start time of this Keyframe.
+   *
    * @return the t1.
    */
-  public int getT1() {
+  public double getT1() {
     return t1;
   }
 
   /**
    * Gets the starting x position of this keyframe.
+   *
    * @return the x1.
    */
-  public int getX1() {
+  public double getX1() {
     return x1;
   }
 
   /**
    * Gets the starting y position of this keyframe.
+   *
    * @return the y1.
    */
-  public int getY1() {
+  public double getY1() {
     return y1;
   }
 
   /**
    * Gets the starting width of the shape in this keyframe.
+   *
    * @return the w1.
    */
-  public int getW1() {
+  public double getW1() {
     return w1;
   }
 
   /**
    * Gets the starting height of the shape in this keyframe.
+   *
    * @return h1.
    */
-  public int getH1() {
+  public double getH1() {
     return h1;
   }
 
   /**
    * Gets the starting r value of the shape in this keyframe.
+   *
    * @return the r1.
    */
-  public int getR1() {
+  public double getR1() {
     return r1;
   }
 
   /**
    * Gets the starting g value of the shape in this keyframe.
+   *
    * @return the g1.
    */
-  public int getG1() {
+  public double getG1() {
     return g1;
   }
 
   /**
    * Gets the starting b value of the shape in this keyframe.
+   *
    * @return the b1.
    */
-  public int getB1() {
+  public double getB1() {
     return b1;
   }
 
   /**
    * Gets the start time of this Keyframe.
+   *
    * @return the t2.
    */
-  public int getT2() {
+  public double getT2() {
     return t2;
   }
 
   /**
    * Gets the starting x position of this keyframe.
+   *
    * @return the x2.
    */
-  public int getX2() {
+  public double getX2() {
     return x2;
   }
 
   /**
    * Gets the starting x position of this keyframe.
+   *
    * @return the y2.
    */
-  public int getY2() {
+  public double getY2() {
     return y2;
   }
 
   /**
    * Gets the starting width of the shape in this keyframe.
+   *
    * @return the w2.
    */
-  public int getW2() {
+  public double getW2() {
     return w2;
   }
 
   /**
    * Gets the starting height of the shape in this keyframe.
+   *
    * @return the h2.
    */
-  public int getH2() {
+  public double getH2() {
     return h2;
   }
 
   /**
    * Gets the starting r value of the shape in this keyframe.
+   *
    * @return the r2.
    */
-  public int getR2() {
+  public double getR2() {
     return r2;
   }
 
   /**
    * Gets the starting g value of the shape in this keyframe.
+   *
    * @return the g2.
    */
-  public int getG2() {
+  public double getG2() {
     return g2;
   }
 
   /**
    * Gets the starting g value of the shape in this keyframe.
+   *
    * @return the g2.
    */
-  public int getB2() {
+  public double getB2() {
     return b2;
   }
+
+
 }
